lzma: Fix issues with 32-bit integers
The type int has 32 bits on 32-bit platforms. Therefore int values
cannot be compared with constants that are larger than the maximum int
value.

+++ lzma/reader.go
@@ -32,7 +32,7 @@ func (c *ReaderConfig) fill() {
 // be replaced by default values.
 func (c *ReaderConfig) Verify() error {
 	c.fill()
-	if !(MinDictCap <= c.DictCap && c.DictCap <= MaxDictCap) {
+	if !(MinDictCap <= c.DictCap && int64(c.DictCap) <= MaxDictCap) {
 		return errors.New("lzma: dictionary capacity is out of range")
 	}
 	return nil
